refactor: finish moving exporting's model and view models
Exporting's services already lived in the feature package, but its
model and view models were still split across the old top-level
folders. Move them in so the slice holds everything the area owns.

QML only has one flat namespace for registered types, so every name
in it needs to carry its own context. Rename the backup dialog and
backup timer view models to carry the Export prefix, matching the
other two view models in the slice that already had it.

The template model relied on a service that still sat in the shared
top-level services layer, which a slice's models role isn't allowed
to reach into directly. Move that service into the slice too.

Move the matching tests into the mirrored test tree, and rename one
of them: it tested the backup timer but was called a backend timer.
