+++ mpvqc/exporting/services/__init__.py
@@ -9,3 +9,5 @@ from .documents import render_classic as render_classic
 from .documents import render_v1 as render_v1
 from .exporter import ExportService as ExportService
 from .settings import ExportSettingsService as ExportSettingsService
+from .template_catalog import ExportTemplate as ExportTemplate
+from .template_catalog import ExportTemplateCatalogService as ExportTemplateCatalogService
