+++ mpvqc/injections.py
@@ -25,7 +25,6 @@ def bindings(binder: inject.Binder) -> None:
     binder.bind_to_constructor(s.CommentTypesPolicyService, s.CommentTypesPolicyService)
     binder.bind_to_constructor(s.CommentTypeValidatorService, s.CommentTypeValidatorService)
     binder.bind_to_constructor(s.DesktopService, s.DesktopService)
-    binder.bind_to_constructor(s.ExportTemplateCatalogService, s.ExportTemplateCatalogService)
     binder.bind_to_constructor(s.FileStartupService, s.FileStartupService)
     binder.bind_to_constructor(s.FontLoaderService, s.FontLoaderService)
     binder.bind_to_constructor(s.InternationalizationService, s.InternationalizationService)
