+++ mpvqc/services/__init__.py
@@ -8,7 +8,6 @@ from .comment_type_validator import CommentTypeValidatorService as CommentTypeVa
 from .comment_types_policy import CommentTypesPolicyService as CommentTypesPolicyService
 from .comments import CommentsService as CommentsService
 from .desktop import DesktopService as DesktopService
-from .export_template_catalog import ExportTemplateCatalogService as ExportTemplateCatalogService
 from .file_startup import FileStartupService as FileStartupService
 from .font_loader import FontLoaderService as FontLoaderService
 from .formatter_time import TimeFormatterService as TimeFormatterService
