+++ mpvqc/viewmodels/__init__.py
@@ -4,7 +4,6 @@
 
 from .app import MpvqcAppViewModel as MpvqcAppViewModel
 from .app import MpvqcResizeViewModel as MpvqcResizeViewModel
-from .file_dialogs import MpvqcExportFileDialogViewModel as MpvqcExportFileDialogViewModel
 from .message_boxes import MpvqcMessageBoxRequestRelayViewModel as MpvqcMessageBoxRequestRelayViewModel
 from .message_boxes import MpvqcQuitMessageBoxViewModel as MpvqcQuitMessageBoxViewModel
 from .message_boxes import MpvqcResetMessageBoxViewModel as MpvqcResetMessageBoxViewModel
@@ -12,7 +11,6 @@ from .message_boxes import MpvqcVersionCheckMessageBoxViewModel as MpvqcVersionC
 from .utility import MpvqcFontsViewModel as MpvqcFontsViewModel
 from .utility import MpvqcLabelWidthCalculatorViewModel as MpvqcLabelWidthCalculatorViewModel
 from .utility import MpvqcWindowViewModel as MpvqcWindowViewModel
-from .views import MpvqcBackupTimerViewModel as MpvqcBackupTimerViewModel
 from .views import MpvqcCommentTableViewModel as MpvqcCommentTableViewModel
 from .views import MpvqcFooterViewModel as MpvqcFooterViewModel
 from .views import MpvqcHeaderViewModel as MpvqcHeaderViewModel
