--- mpvqc/viewmodels/file_dialogs/__init__.py
@@ -1,5 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from .export_file_dialog import MpvqcExportFileDialogViewModel as MpvqcExportFileDialogViewModel
