+++ mpvqc/viewmodels/views/table/__init__.py
@@ -2,7 +2,6 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .backup_timer import MpvqcBackupTimerViewModel as MpvqcBackupTimerViewModel
 from .comment_table import MpvqcCommentTableViewModel as MpvqcCommentTableViewModel
 from .placeholder import MpvqcPlaceholderViewModel as MpvqcPlaceholderViewModel
 from .search_box import MpvqcSearchBoxViewModel as MpvqcSearchBoxViewModel
