+++ pyproject.toml
@@ -142,11 +142,9 @@ files = [
     "mpvqc/close_event_filter.py",
     "mpvqc/dialogs/__init__.py",
     "mpvqc/dialogs/about.py",
-    "mpvqc/dialogs/backup.py",
     "mpvqc/dialogs/comment_types.py",
     "mpvqc/dialogs/edit_input.py",
     "mpvqc/dialogs/edit_mpv.py",
-    "mpvqc/dialogs/export_settings.py",
     "mpvqc/enums/__init__.py",
     "mpvqc/enums/dialog_kind.py",
     "mpvqc/enums/file_dialog_kind.py",
@@ -154,6 +152,8 @@ files = [
     "mpvqc/enums/time_display_mode.py",
     "mpvqc/enums/window_title_format.py",
     "mpvqc/exporting/__init__.py",
+    "mpvqc/exporting/models/__init__.py",
+    "mpvqc/exporting/models/templates.py",
     "mpvqc/exporting/services/__init__.py",
     "mpvqc/exporting/services/backup.py",
     "mpvqc/exporting/services/context.py",
@@ -162,7 +162,13 @@ files = [
     "mpvqc/exporting/services/documents/v1.py",
     "mpvqc/exporting/services/exporter.py",
     "mpvqc/exporting/services/settings.py",
+    "mpvqc/exporting/services/template_catalog.py",
     "mpvqc/exporting/services/writer.py",
+    "mpvqc/exporting/viewmodels/__init__.py",
+    "mpvqc/exporting/viewmodels/backup_dialog.py",
+    "mpvqc/exporting/viewmodels/backup_timer.py",
+    "mpvqc/exporting/viewmodels/export_settings_dialog.py",
+    "mpvqc/exporting/viewmodels/file_dialog.py",
     "mpvqc/exporting/wiring.py",
     "mpvqc/importing/__init__.py",
     "mpvqc/importing/domain/__init__.py",
@@ -202,7 +208,6 @@ files = [
     "mpvqc/logging_utils.py",
     "mpvqc/models/__init__.py",
     "mpvqc/models/dependencies.py",
-    "mpvqc/models/export_templates.py",
     "mpvqc/models/languages.py",
     "mpvqc/models/shortcuts.py",
     "mpvqc/services/__init__.py",
@@ -220,7 +225,6 @@ files = [
     "mpvqc/services/comments/store.py",
     "mpvqc/services/comments/view_action.py",
     "mpvqc/services/desktop.py",
-    "mpvqc/services/export_template_catalog.py",
     "mpvqc/services/file_startup.py",
     "mpvqc/services/font_loader.py",
     "mpvqc/services/formatter_time.py",
@@ -273,8 +277,6 @@ files = [
     "mpvqc/viewmodels/app/__init__.py",
     "mpvqc/viewmodels/app/app.py",
     "mpvqc/viewmodels/app/resize.py",
-    "mpvqc/viewmodels/file_dialogs/__init__.py",
-    "mpvqc/viewmodels/file_dialogs/export_file_dialog.py",
     "mpvqc/viewmodels/message_boxes/__init__.py",
     "mpvqc/viewmodels/message_boxes/quit.py",
     "mpvqc/viewmodels/message_boxes/request_relay.py",
@@ -295,7 +297,6 @@ files = [
     "mpvqc/viewmodels/views/player/new_comment_menu.py",
     "mpvqc/viewmodels/views/player/player.py",
     "mpvqc/viewmodels/views/table/__init__.py",
-    "mpvqc/viewmodels/views/table/backup_timer.py",
     "mpvqc/viewmodels/views/table/comment_table.py",
     "mpvqc/viewmodels/views/table/placeholder.py",
     "mpvqc/viewmodels/views/table/search_box.py",
